Varying LED blink rate as distance gets lower
Added timer function to call Sense_distance function where we call blink function when distance < 200 cm. we pass rate as param to blink led function so it blinks faster as distance gets low and stays solid when is <= 30 cm. Light stays off when distance > 200 cm. 
Things to fix in this version are
- distance sensor is very sensitive and measurements vary drastically. Need hw/sw changes to get that in control
- add door status to include that in the condition to call blink led function.